requirements.txt
asgiref==3.4.1
autopep8==1.6.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cleanup==5.2.0
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
firebase-admin==5.2.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
jws==0.1.3
Markdown==3.3.6
msgpack==1.0.3
oauth2client==3.0.0
packaging==21.3
Pillow==9.0.0
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.7
python-jwt==2.0.1
pytz==2021.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
